+++ crates/loro-internal/examples/encoding.rs
@@ -62,6 +62,15 @@ fn main() {
         output.len(),
     );
 
+    let json_updates =
+        serde_json::to_string(&loro.export_json_updates(&Default::default())).unwrap();
+    let output = miniz_oxide::deflate::compress_to_vec(json_updates.as_bytes(), 6);
+    println!(
+        "json updates size {} after compression {}",
+        json_updates.len(),
+        output.len(),
+    );
+
     // {
     //     // Delta encoding