+++ crates/loro-internal/src/state/tree_state.rs
@@ -972,15 +972,31 @@ impl ContainerState for TreeState {
 
     fn apply_local_op(&mut self, raw_op: &RawOp, _op: &Op) -> LoroResult<()> {
         match &raw_op.content {
-            crate::op::RawOpContent::Tree(tree) => {
-                let TreeOp {
+            crate::op::RawOpContent::Tree(tree) => match tree {
+                TreeOp::Create {
                     target,
                     parent,
                     position,
-                } = tree;
-                let parent = TreeParentId::from(*parent);
-                self.mov(*target, parent, raw_op.id_full(), position.clone(), true)
-            }
+                }
+                | TreeOp::Move {
+                    target,
+                    parent,
+                    position,
+                } => {
+                    let parent = TreeParentId::from(*parent);
+                    self.mov(
+                        *target,
+                        parent,
+                        raw_op.id_full(),
+                        Some(position.clone()),
+                        true,
+                    )
+                }
+                TreeOp::Delete { target } => {
+                    let parent = TreeParentId::Deleted;
+                    self.mov(*target, parent, raw_op.id_full(), None, true)
+                }
+            },
             _ => unreachable!(),
         }
     }
@@ -1080,12 +1096,27 @@ impl ContainerState for TreeState {
         for op in ctx.ops {
             assert_eq!(op.op.atom_len(), 1);
             let content = op.op.content.as_tree().unwrap();
-            let target = content.target;
-            let parent = content.parent;
-            let position = content.position.clone();
-            let parent = TreeParentId::from(parent);
-            self.mov(target, parent, op.id_full(), position, false)
-                .unwrap();
+            match content {
+                TreeOp::Create {
+                    target,
+                    parent,
+                    position,
+                }
+                | TreeOp::Move {
+                    target,
+                    parent,
+                    position,
+                } => {
+                    let parent = TreeParentId::from(*parent);
+                    self.mov(*target, parent, op.id_full(), Some(position.clone()), false)
+                        .unwrap()
+                }
+                TreeOp::Delete { target } => {
+                    let parent = TreeParentId::Deleted;
+                    self.mov(*target, parent, op.id_full(), None, false)
+                        .unwrap()
+                }
+            };
         }
     }
 }